refactor(src/app): Refactor App component
- Move logic, template and styling of App Component into a new component
called List. App Component should have minimum responsibility. Hence,
List should be solely responsible for managing the todo items.
