--- src/app/app.component.css
@@ -1,136 +1,0 @@
-body {
-  color: #4d4d4d;
-  background-color: #f5f5f5;
-  color: #4d4d4d;
-}
-
-p {
-  font-size: 1.2rem;
-}
-
-.main {
-  max-width: 500px;
-  width: 85%;
-  margin: 2rem auto;
-  padding: 1rem;
-  text-align: center;
-  box-shadow: 0 2px 4px 0 hsla(0, 0%, 0%, 0.2), 0 2.5rem 5rem 0 hsla(0, 0%, 0%, 0.1);
-}
-
-@media screen and (min-width: 600px) {
-  .main {
-    width: 70%;
-  }
-}
-
-label {
-  font-size: 1.5rem;
-  font-weight: bold;
-  display: block;
-  padding-bottom: 1rem;
-}
-
-.lg-text-input {
-  width: 100%;
-  padding: 1rem;
-  border: 2px solid #000;
-  display: block;
-  box-sizing: border-box;
-  font-size: 1rem;
-}
-
-a {
-  text-decoration: none;
-  color: #45a519;
-}
-
-a:hover {
-  color: red;
-}
-
-.emoji {
-  font-size: 2rem;
-  vertical-align: middle;
-}
-
-.btn-wrapper {
-  margin-bottom: 2rem;
-}
-
-.btn-menu {
-  flex-basis: 32%;
-}
-
-.active {
-  color: green;
-}
-
-ul {
-  padding-inline-start: 0;
-}
-
-ul li {
-  list-style: none;
-}
-
-.main-dark {
-  background-color: #15202B;
-  color: #fff;
-  box-shadow: 0 2px 4px 0 hsla(0, 6%, 97%, 0.2);
-}
-
-.header {
-  display: flex;
-  align-items: center;
-}
-
-.btn-ghost {
-  margin-left: auto;
-  margin-bottom: auto;
-  cursor: pointer;
-}
-
-ul.itemsList {
-  padding-left: 15px;
-}
-
-.card {
-  display: inline-block;
-  padding: 2rem 1rem 5rem 1rem;
-  background-color: lemonchiffon;
-  position: relative;
-  border: none;
-  color: #000;
-}
-
-.card:before {
-  content: "";
-  display: block;
-  width: 0px;
-  height: 0px;
-  border-top: 30px solid lemonchiffon;
-  border-right: 30px solid lemonchiffon;
-  border-left: 30px solid lightblue;
-  border-bottom: 30px solid lightblue;
-  bottom: 0px;
-  left: 0px;
-  position: absolute;
-  filter: drop-shadow(7px -7px 5px #706e6e);
-  margin-right: 10%;
-}
-
-.card:after {
-  content: "";
-  display: block;
-  width: 0px;
-  height: 0px;
-  background-color: skyblue;
-  top: 0px;
-  right: 0px;
-  border-bottom: 20px solid lemonchiffon;
-  border-left: 20px solid lemonchiffon;
-  border-right: 20px solid lightblue;
-  border-top: 20px solid lightblue;
-  position: absolute;
-  filter: drop-shadow(-5px 5px 3px #706e6e);
-}
